+++ core/src/main/java/Gameplay/Movimientos/Movimiento.java
@@ -9,20 +9,14 @@ public abstract class Movimiento implements IMovimiento {
     protected String nombre;
     protected Texture textura;
     protected Sprite sprite;
-    protected int cooldown;
 
-    public Movimiento(String nombre, Texture textura, int cooldown) {
+    public Movimiento(String nombre, Texture textura) {
         this.nombre = nombre;
         this.textura = textura;
         this.sprite = new Sprite(textura);
-        this.cooldown = cooldown;
     }
-
-    public abstract void ejecutar(Personaje personaje, float potencia);
 
     public String getNombre() { return nombre; }
     public Sprite getSprite() { return sprite; }
-    public int getCooldown() { return cooldown; }
-    public void setCooldown(int cooldown) { this.cooldown = cooldown; }
 
 }