+++ core/src/main/java/entidades/Limite.java
@@ -6,7 +6,9 @@ import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 import com.badlogic.gdx.math.Rectangle;
 
 public class Limite implements Colisionable {
+
     private Rectangle hitbox;
+    private boolean activo = true;
 
     public Limite(float x, float y, float ancho, float alto) {
         this.hitbox = new Rectangle(x, y, ancho, alto);
@@ -28,28 +30,33 @@ public class Limite implements Colisionable {
     }
 
     @Override
-    public void desactivar() {}
+    public void desactivar() {
+        activo = false;
+    }
 
     @Override
     public float getX() {
-        return 0;
+        return hitbox.x;
     }
 
     @Override
     public float getY() {
-        return 0;
+        return hitbox.y;
     }
 
     @Override
     public void setX(float x) {
-
+        hitbox.x = x;
     }
 
     @Override
     public void setY(float y) {
-
+        hitbox.y = y;
     }
 
     @Override
-    public boolean getActivo() { return false; }
+    public boolean getActivo() {
+        return activo;
+    }
 }
+