+++ core/src/main/java/entidades/personajes/HormigaExploradora.java
@@ -2,6 +2,7 @@ package entidades.personajes;
 
 import Gameplay.Gestores.GestorColisiones;
 import Gameplay.Gestores.GestorProyectiles;
+import Gameplay.Movimientos.Melee.Aranazo;
 import Gameplay.Movimientos.Rango.LanzaGranada;
 import Gameplay.Movimientos.Rango.LanzaNectar;
 import Gameplay.Movimientos.Rango.LanzaRoca;
@@ -17,10 +18,10 @@ public class HormigaExploradora extends Personaje {
 
     @Override
     protected void inicializarMovimientos() {
-        Texture texturaMovimiento = GestorAssets.get(Constantes.MAPA_1, Texture.class);
-        movimientos.add(new LanzaRoca("Lanzar Roca", texturaMovimiento, 1, 300f, gestorProyectiles));
-        movimientos.add(new LanzaNectar("Lanza Nectar", texturaMovimiento, 1, 300f, gestorProyectiles));
-        movimientos.add(new LanzaGranada("Lanza Granada", texturaMovimiento, 1, 300f, gestorProyectiles));
+        movimientos.add(new Aranazo(gestorColisiones));
+        movimientos.add(new LanzaNectar(gestorProyectiles));
+        movimientos.add(new LanzaGranada(gestorProyectiles));
+        movimientos.add(new LanzaRoca(gestorProyectiles));
 
     }