+++ core/src/main/java/entidades/personajes/HormigaObrera.java
@@ -2,6 +2,9 @@ package entidades.personajes;
 
 import Gameplay.Gestores.GestorColisiones;
 import Gameplay.Gestores.GestorProyectiles;
+import Gameplay.Movimientos.Melee.Aranazo;
+import Gameplay.Movimientos.Rango.LanzaGranada;
+import Gameplay.Movimientos.Rango.LanzaNectar;
 import Gameplay.Movimientos.Rango.LanzaRoca;
 import com.badlogic.gdx.graphics.Texture;
 import managers.GestorAssets;
@@ -15,8 +18,10 @@ public class HormigaObrera extends Personaje {
 
     @Override
     protected void inicializarMovimientos() {
-        Texture texturaMovimiento = GestorAssets.get(Constantes.PNG_1, Texture.class);
-        movimientos.add(new LanzaRoca("Lanzar Roca", texturaMovimiento, 1, 300f, gestorProyectiles));
+        movimientos.add(new Aranazo(gestorColisiones));
+        movimientos.add(new LanzaNectar(gestorProyectiles));
+        movimientos.add(new LanzaGranada(gestorProyectiles));
+        movimientos.add(new LanzaRoca(gestorProyectiles));
 
     }